requirements.txt
alembic==1.6.5
cachelib==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
dataclasses==0.8
Faker==8.10.1
Flask==1.1.4
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Moment==1.0.2
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-Share==0.1.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.19.5
opencv-python==4.4.0
Pillow==8.3.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.4
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
